+++ cocos/scene/Model.h
@@ -79,7 +79,7 @@ public:
         }
         *_worldBounds = aabb;
     }
-    inline void setInstancedAttrBlock(uint8_t *buffer, size_t size, InstancedAttributeBlock &&block, const std::vector<gfx::Attribute> &attributes) {
+    inline void setInstancedAttrBlock(uint8_t *buffer, uint32_t size, InstancedAttributeBlock &&block, const std::vector<gfx::Attribute> &attributes) {
         _instancedBuffer        = {buffer, size};
         _instanceAttributeBlock = std::move(block);
         _instanceAttributes     = attributes;
@@ -91,7 +91,7 @@ public:
     inline const std::vector<gfx::Attribute> &getInstanceAttributes() const { return _instanceAttributes; }
     inline InstancedAttributeBlock *          getInstancedAttributeBlock() { return &_instanceAttributeBlock; }
     inline uint8_t *                          getInstancedBuffer() const { return std::get<0>(_instancedBuffer); }
-    inline size_t                             getInstancedBufferSize() const { return std::get<1>(_instancedBuffer); }
+    inline uint32_t                           getInstancedBufferSize() const { return std::get<1>(_instancedBuffer); }
     inline gfx::Buffer *                      getLocalBuffer() const { return _localBuffer; }
     inline float *                            getLocalData() const { return _localData; }
     inline const AABB &                       getModelBounds() const { return _modelBounds; }
@@ -116,18 +116,18 @@ private:
     bool _castShadow{false};
     bool _receiveShadow{false};
 
-    int                           _instmatWorldIdx{0};
-    uint32_t                      _visFlags;
-    uint32_t                      _updateStamp{0};
-    Node *                        _transform{nullptr};
-    Node *                        _node{nullptr};
-    float *                       _localData{nullptr};
-    std::tuple<uint8_t *, size_t> _instancedBuffer{nullptr, 0};
-    gfx::Buffer *                 _localBuffer{nullptr};
-    InstancedAttributeBlock       _instanceAttributeBlock{};
-    std::vector<SubModel *>       _subModels;
-    std::vector<gfx::Attribute>   _instanceAttributes;
-    static void                   uploadMat4AsVec4x3(const Mat4 &mat, float *v1, float *v2, float *v3);
+    int                             _instmatWorldIdx{0};
+    uint32_t                        _visFlags;
+    uint32_t                        _updateStamp{0};
+    Node *                          _transform{nullptr};
+    Node *                          _node{nullptr};
+    float *                         _localData{nullptr};
+    std::tuple<uint8_t *, uint32_t> _instancedBuffer{nullptr, 0};
+    gfx::Buffer *                   _localBuffer{nullptr};
+    InstancedAttributeBlock         _instanceAttributeBlock{};
+    std::vector<SubModel *>         _subModels;
+    std::vector<gfx::Attribute>     _instanceAttributes;
+    static void                     uploadMat4AsVec4x3(const Mat4 &mat, float *v1, float *v2, float *v3);
 };
 
 } // namespace scene
